+++ packages/camera/camera/lib/src/camera_controller.dart
@@ -24,8 +24,12 @@ typedef onLatestImageAvailable = void Function(CameraImage image);
 /// Completes with a list of available cameras.
 ///
 /// May throw a [CameraException].
-Future<List<CameraDescription>> availableCameras() async {
-  return CameraPlatform.instance.availableCameras();
+Future<List<CameraDescription>> availableCameras({
+  bool physicalCameras = true,
+  bool logicalCameras = false,
+}) async {
+  return CameraPlatform.instance.availableCameras(
+      logicalCameras: logicalCameras, physicalCameras: physicalCameras);
 }
 
 // TODO(stuartmorgan): Remove this once the package requires 2.10, where the