+++ packages/camera/camera/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Updates support matrix in README to indicate that iOS 11 is no longer supported.
 * Clients on versions of Flutter that still support iOS 11 can continue to use this
   package with iOS 11, but will not receive any further updates to the iOS implementation.
+* Update availableDevices method to specify if it will return physical and/or logical devices.
 
 ## 0.10.5+9