+++ packages/camera/camera_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Add placeholder physicalCameras and logicalCameras options to availableDevices method.
+
 ## 0.3.2+4
 
 * Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.